✨: change double to float

+++ Calculator/app/src/main/java/com/example/calculator_oop_kotlin_project/HexadecimalActivity.kt
@@ -19,7 +19,6 @@ class HexadecimalActivity : AppCompatActivity() {
         setOnClickListeners()
     }
 
-
     private fun setOnClickListeners()
     {
         options_btn.setOnClickListener { options() }
@@ -85,23 +84,23 @@ class HexadecimalActivity : AppCompatActivity() {
 
         lastOperator = OperatorClass
 
-        var result:Double=0.0
+        var result:Float=0.0f
 
         var tempValue1 = firstNumber.text.toString()
         var tempValue2 = secondNumber.text.toString()
 
         when {
             !firstNumber.text.isNullOrEmpty()  && secondNumber.text.isNullOrEmpty() -> {
-                result = tempValue1.toDouble()
+                result = tempValue1.toFloat()
             }
 
             (firstNumber.text=="0.0" || firstNumber.text=="-0.0" ) && !secondNumber.text.isNullOrEmpty() -> {
-                result = tempValue2.toDouble()
+                result = tempValue2.toFloat()
             }
 
             !firstNumber.text.isNullOrEmpty() && !secondNumber.text.isNullOrEmpty() ->
             {
-                result = OperatorClass.checkTwoNumbers(tempValue1.toDouble(), tempValue2.toDouble())
+                result = OperatorClass.checkTwoNumbers(tempValue1.toFloat(), tempValue2.toFloat())
             }
 
             firstNumber.text.isNullOrEmpty() && !secondNumber.text.isNullOrEmpty() -> {
@@ -111,7 +110,7 @@ class HexadecimalActivity : AppCompatActivity() {
             }
 
             !firstNumber.text.isNullOrEmpty() ->{
-                result = OperatorClass.checkOneNumber(tempValue1.toDouble())
+                result = OperatorClass.checkOneNumber(tempValue1.toFloat())
             }
         }
